Change the default phase 2 reconstruction for endcap from simPF to pfTICL.
